+++ SangmyungBank/src/main/java/com/EmperorPenguin/SangmyungBank/api/securitynotices/domain/securitynotices/SecurityNotices.java
@@ -1,6 +1,5 @@
 package com.EmperorPenguin.SangmyungBank.api.securitynotices.domain.securitynotices;
 
-import com.EmperorPenguin.SangmyungBank.api.users.add.domain.User.User;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
