+++ SangmyungBank/src/test/java/com/EmperorPenguin/SangmyungBank/api/user/login/loginUserTest.java
@@ -1,6 +1,6 @@
 package com.EmperorPenguin.SangmyungBank.api.user.login;
 
-import com.EmperorPenguin.SangmyungBank.api.users.add.domain.User.User;
+import com.EmperorPenguin.SangmyungBank.api.users.register.domain.User.User;
 import com.EmperorPenguin.SangmyungBank.api.users.login.domain.loginForm.LoginForm;
 import com.EmperorPenguin.SangmyungBank.api.users.login.service.UserLoginService;
 import org.junit.jupiter.api.Test;
